requirements.txt
arxiv==2.4.1
PyMuPDF==1.27.2.2
transformers==5.5.3
peft==0.19.1
bitsandbytes==0.49.2
accelerate==1.13.0
datasets==4.8.4
pyyaml==6.0.3
fastapi==0.134.0
uvicorn==0.32.1
torch-geometric==2.7.0
pandas==2.3.3
scikit-learn==1.8.0
